Bumped to 2.1.0
Printing Hex will now pay attention to the width parameter better than
std::ostream does. Width now specifies the number of digits in the printed
result (but not the actual width in characters. The 0x prefix adds two more,
and if the separater is not null then we'd add log_64k (N) separators.